requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==4.0.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
mccabe==0.6.1
oauthlib==3.2.0
protobuf==4.21.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.39
uritemplate==4.1.1
urllib3==1.26.11
xmltodict==0.13.0
